requirements.txt
glob2==0.6
h5py==2.8.0
imageio==2.3.0
lmdb==0.98
numba==0.38.0
numpy==1.16.4
numpydoc==0.8.0
pandas==0.24.2
pathlib2==2.3.2
pdbpp==0.10.2
pep8==1.7.1
Pillow==5.4.1
pkginfo==1.4.2
protobuf==3.8.0
protobuf3-to-dict==0.1.5
PyYAML==3.12
requests==2.23.0
requests-kerberos==0.12.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.87
scikit-image==0.13.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.8.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
torch==1.5.0
torchvision==0.6.0
tornado==5.0.2
tqdm==4.45.0
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zict==0.1.3
zipp==3.0.0
